# Conflicts:
#	app/Traits/CalendarDateTrait.php
#	public/css/app.css
#	public/js/abilities.js
#	public/js/api.js
#	public/js/app.js
#	public/js/attributes.js
#	public/js/community-votes.js
#	public/js/conversation.js
#	public/mix-manifest.json
#	resources/views/quests/characters/_form.blade.php
#	resources/views/quests/items/_form.blade.php
#	resources/views/quests/locations/_form.blade.php
#	resources/views/quests/organisations/_form.blade.php

+++ app/Observers/EntityObserver.php
@@ -238,6 +238,11 @@ class EntityObserver
      */
     public function deleted(Entity $entity)
     {
+        // If soft deleting, don't really delete the image
+        if ($entity->trashed()) {
+            return;
+        }
+
         $entity->permissions()->delete();
         $entity->widgets()->delete();
     }
